refactor: Moving Metrics to Observability
